+++ src/base/logging/sinks/stdout.cpp
@@ -32,7 +32,7 @@ namespace Base
 				return;
 			std::wstring UTF16Message = Details::UTF8ToUTF16(message);
 			DWORD numberOfCharsWritten;
-			WriteConsoleW(_stdhandle, UTF16Message.data(), UTF16Message.size(), &numberOfCharsWritten, NULL);
+			WriteConsoleW(_stdhandle, UTF16Message.data(), DWORD(UTF16Message.size()), &numberOfCharsWritten, NULL);
 		}
 
 		void StdOutSink::flush()
